+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [UNRELEASED]
+
+- Turn off linter and test coverage for generated files
+
 ## 1.3.0 - 2025-11-25
 
 - Add `.map(elementTransformer)` for ZArray